pyproject.toml
[project]
name = "fantasy-golf-backend"
version = "0.1.0"
description = "League Caddie Backend API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.14.0",
    "psycopg2-binary>=2.9.9",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    "httpx>=0.27.0",
    "apscheduler>=3.10.0",
    "pydantic-settings>=2.5.0",
    "google-auth>=2.35.0",
    "requests>=2.32.0",
    "slowapi>=0.1.9",
    "boto3>=1.34.0",
    "email-validator>=2.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.7.0",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
